backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy==1.4.49",
    "uvicorn[standard]>=0.35.0",
]
